# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cloud IDE API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List all files in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: Created file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
  /projects/{id}/files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get a specific file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
    patch:
      operationId: updateFile
      tags: [files]
      summary: Update a file's content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        template:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived]
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - language
        - status
        - isPublic
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string
        template:
          type: string
        isPublic:
          type: boolean
      required:
        - name
        - language
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, archived]
        isPublic:
          type: boolean
...
    File:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        path:
          type: string
        content:
          type: ["string", "null"]
        type:
          type: string
          enum: [file, directory]
        language:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - name
        - path
        - type
        - createdAt
        - updatedAt
    FileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        path:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [file, directory]
      required:
        - name
        - path
        - type
    FileUpdate:
      type: object
      properties:
        content:
          type: string
        name:
          type: string
...