# lib/api-zod/src/generated/types/file.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud IDE API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileType } from './fileType';

export interface File {
  id: number;
  projectId: number;
  name: string;
  path: string;
  /** @nullable */
  content?: string | null;
  type: FileType;
  /** @nullable */
  language?: string | null;
  createdAt: string;
  updatedAt: string;
}